+++ app/Console/Kernel.php
@@ -1,5 +1,6 @@
 <?php namespace App\Console;
 
+use App\Jobs\HarvestJob;
 use Illuminate\Console\Scheduling\Schedule;
 use Illuminate\Foundation\Console\Kernel as ConsoleKernel;
 
@@ -38,12 +39,12 @@ class Kernel extends ConsoleKernel
         $schedule->command('sitemap:make')->daily();
         $schedule->call(function() {
             foreach (\App\SpiceHarvesterHarvest::where('cron_status', '=', 'daily')->get() as $harvest) {
-                App::make('\App\Http\Controllers\SpiceHarvesterController')->launch($harvest->id);
+                dispatch(new HarvestJob($harvest));
             }
         })->daily(); /* daily at midnight */
         $schedule->call(function() {
             foreach (\App\SpiceHarvesterHarvest::where('cron_status', '=', 'weekly')->get() as $harvest) {
-                App::make('\App\Http\Controllers\SpiceHarvesterController')->launch($harvest->id);
+                dispatch(new HarvestJob($harvest));
             }
         })->weeklyOn(6, '23:00'); /* sundays at 11pm */
     }