+++ app/Jobs/HarvestJob.php
@@ -35,6 +35,9 @@ class HarvestJob extends Job implements ShouldQueue
         $this->to = $to;
         $this->only_ids = $only_ids;
 
+        $harvest->status = SpiceHarvesterHarvest::STATUS_QUEUED;
+        $harvest->save();
+
     }
 
     public function handle(SpiceHarvesterService $spiceHarvesterService) {