+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file[^1].
 - missing relations between items and authorities in harvest
 - syncing relations in harvest
 - failing harvests
+- scheduled harvest jobs
 - harvest only alternative authority names
 
 ### Added